agent/requirements.txt
google-adk>=1.0.0
litellm>=1.75.5
httpx>=0.27.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
a2a-sdk>=0.2.0,<1.0.0
